ホストのロール割り当てログが割り当て前だったので表示タイミングを修正
